requirements.txt
absl-py==2.1.0
accelerate==0.29.3
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.5.0
anyio==4.9.0
argbind==0.3.9
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
bert-score==0.3.13
bitsandbytes==0.43.1
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
datasets==2.19.0
decorator==5.1.1
deepspeed==0.9.5
descript-audio-codec==1.0.0
descript-audiotools==0.7.2
dill==0.3.8
docstring_parser==0.16
einops==0.8.0
evaluate==0.4.3
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.115.12
ffmpy==0.3.2
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
flatten-dict==0.4.2
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
gradio==3.50.2
gradio_client==0.6.1
graphviz==0.20.3
grpcio==1.64.1
h11==0.16.0
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
julius==0.2.7
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.5.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
narwhals==1.40.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
openai==1.82.0
orjson==3.10.18
packaging==24.1
pandas==2.2.2
parso==0.8.4
peft==0.10.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.2
prompt_toolkit==3.0.47
protobuf==3.19.6
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycparser==2.22
pydantic==1.10.17
pydub==0.25.1
Pygments==2.18.0
pyloudnorm==0.1.1
pyparsing==3.1.2
pystoi==0.4.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.1
randomname==0.2.1
referencing==0.36.2
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rouge_score==0.1.2
rpds-py==0.25.0
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.13.0
semantic-version==2.10.0
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.3.7
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1
torch-stoi==0.2.1
torchaudio==2.3.1
torchviz==0.0.2
tqdm==4.66.4
traitlets==5.14.3
transformers==4.40.1
triton==2.3.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.34.2
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.3
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
